+++ docs/source/conf.py
@@ -7,8 +7,7 @@
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = "molflux"
-# copyright = "2023, Exscientia"
-# author = "Exscientia"
+author = "Exscientia"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration